src/backend/requirements.txt
Flask==2.2.2
pymongo==4.2.0
Authlib==1.1.0
requests==2.28.1
python-dotenv==0.19.2
tweepy==4.10.1
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
py-trello==0.18.0
PyGithub==1.56
spotipy==2.21.0
